MainWeb_Final/AI_asset/requirements.txt
flask
flask-cors
torch
torchvision
pillow
pandas
scikit-learn
